[ADD] 댓글 등록, 이모지 댓글 등록 param validation isNumeric으로 수정, 댓글 삭제 validation 추가 #90

+++ src/controller/commentController.ts
@@ -12,6 +12,7 @@ const createComment = async (
   next: NextFunction
 ) => {
   const userId: number = req.body.userId;
+
   try {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -79,6 +80,13 @@ const createEmojiComment = async (
 
 //? 댓글 삭제하기
 const deleteComment = async (req: Request, res: Response) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res
+      .status(sc.BAD_REQUEST)
+      .send(fail(sc.BAD_REQUEST, rm.BAD_REQUEST));
+  }
+
   const { commentId } = req.params;
 
   await commentService.deleteComment(+commentId);
